+++ crates/sequencer/src/state.rs
@@ -17,6 +17,7 @@ use starknet_api::{
 use serde::{Deserialize, Serialize};
 
 use crate::commit::Committer;
+use crate::serde::SerializableState;
 
 /// Generic state structure for the sequencer.
 /// The use of `FxHashMap` allows for a better performance.
@@ -26,11 +27,35 @@ use crate::commit::Committer;
 /// See [rustc-hash](https://crates.io/crates/rustc-hash) for more information.
 #[derive(Clone, Default, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct State {
-    pub(crate) classes: FxHashMap<ClassHash, ContractClass>,
-    pub(crate) compiled_class_hashes: FxHashMap<ClassHash, CompiledClassHash>,
-    pub(crate) contracts: FxHashMap<ContractAddress, ClassHash>,
-    pub(crate) storage: FxHashMap<ContractStorageKey, StarkFelt>,
-    pub(crate) nonces: FxHashMap<ContractAddress, Nonce>,
+    classes: FxHashMap<ClassHash, ContractClass>,
+    compiled_class_hashes: FxHashMap<ClassHash, CompiledClassHash>,
+    contracts: FxHashMap<ContractAddress, ClassHash>,
+    storage: FxHashMap<ContractStorageKey, StarkFelt>,
+    nonces: FxHashMap<ContractAddress, Nonce>,
+}
+
+impl From<State> for SerializableState {
+    fn from(state: State) -> Self {
+        Self {
+            classes: state.classes,
+            compiled_classes_hash: state.compiled_class_hashes,
+            contracts: state.contracts,
+            storage: state.storage,
+            nonces: state.nonces,
+        }
+    }
+}
+
+impl From<SerializableState> for State {
+    fn from(serializable_state: SerializableState) -> Self {
+        Self {
+            classes: serializable_state.classes,
+            compiled_class_hashes: serializable_state.compiled_classes_hash,
+            contracts: serializable_state.contracts,
+            storage: serializable_state.storage,
+            nonces: serializable_state.nonces,
+        }
+    }
 }
 
 impl State {
